[REFACTOR] : 타입스크립트 전환
