+++ src/components/InputTodo.tsx
@@ -1,0 +1,69 @@
+import { FaPlusCircle, FaSpinner } from 'react-icons/fa';
+import { useCallback, useEffect, useState } from 'react';
+
+import { createTodo } from '../api/todo';
+import useFocus from '../hooks/useFocus';
+
+type InputTodoProps = {
+	setTodos: React.Dispatch<React.SetStateAction<Todo[]>>,
+};
+
+const InputTodo = ({ setTodos }: InputTodoProps) => {
+	const [inputText, setInputText] = useState('');
+	const [isLoading, setIsLoading] = useState(false);
+	const { ref, setFocus } = useFocus();
+
+	useEffect(() => {
+		setFocus();
+	}, [setFocus]);
+
+	const handleSubmit = useCallback(
+		async (e: React.FormEvent<HTMLFormElement>) => {
+			try {
+				e.preventDefault();
+				setIsLoading(true);
+
+				const trimmed = inputText.trim();
+				if (!trimmed) {
+					return alert('Please write something');
+				}
+
+				const newItem = { title: trimmed };
+				const { data } = await createTodo(newItem);
+
+				if (data) {
+					return setTodos((prev) => [...prev, data]);
+				}
+			} catch (error) {
+				console.error(error);
+				alert('Something went wrong.');
+			} finally {
+				setInputText('');
+				setIsLoading(false);
+			}
+		},
+		[inputText, setTodos]
+	);
+
+	return (
+		<form className="form-container" onSubmit={handleSubmit}>
+			<input
+				className="input-text"
+				placeholder="Add new todo..."
+				ref={ref}
+				value={inputText}
+				onChange={(e) => setInputText(e.target.value)}
+				disabled={isLoading}
+			/>
+			{!isLoading ? (
+				<button className="input-submit" type="submit">
+					<FaPlusCircle className="btn-plus" />
+				</button>
+			) : (
+				<FaSpinner className="spinner" />
+			)}
+		</form>
+	);
+};
+
+export default InputTodo;
