--- src/components/TodoList.js
@@ -1,14 +1,0 @@
-import TodoItem from "./TodoItem";
-
-const TodoList = ({ todos, setTodos }) => {
-  return todos.length ? (
-    <ul>
-      {todos.map(({ id, title }) => (
-        <TodoItem key={id} id={id} title={title} setTodos={setTodos} />
-      ))}
-    </ul>
-  ) : (
-    <div className="empty-list">...</div>
-  );
-};
-export default TodoList;
