+++ src/components/TodoList.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import TodoItem from './TodoItem';
+
+type TodoListProps = {
+	todos: Todo[],
+	setTodos: React.Dispatch<React.SetStateAction<Todo[]>>,
+};
+
+const TodoList = ({ todos, setTodos }: TodoListProps) => {
+	return todos.length ? (
+		<ul>
+			{todos.map(({ id, title }) => (
+				<TodoItem key={id} id={id} title={title} setTodos={setTodos} />
+			))}
+		</ul>
+	) : (
+		<div className="empty-list">...</div>
+	);
+};
+export default TodoList;
