--- src/hooks/useFocus.js
@@ -1,12 +1,0 @@
-import { useRef } from "react";
-
-const useFocus = () => {
-  const ref = useRef();
-  const setFocus = () => {
-    ref.current && ref.current.focus();
-  };
-
-  return { ref, setFocus };
-};
-
-export default useFocus;
