+++ src/index.tsx
@@ -1,0 +1,10 @@
+import App from './App';
+import ReactDOM from 'react-dom';
+import { StrictMode } from 'react';
+
+ReactDOM.render(
+	<StrictMode>
+		<App />
+	</StrictMode>,
+	document.getElementById('root')
+);
