+++ src/types/todo.d.ts
@@ -1,0 +1,6 @@
+declare type Todo = {
+	createdAt: string;
+	id: string;
+	title: string;
+	updatedAt: string;
+};
